refactor(bin-inspector): split useBinInspector into focused modules
Decompose the 663-LOC useBinInspector.ts into a slim main hook plus
two focused sibling modules. No behavior change; public hook contract
preserved (40 existing tests still pass).

- useBinInspectorTypes.ts (96 LOC): BinField/BinConstraints/
  ConfirmDeleteState/UseBinInspectorReturn interfaces +
  emitLinkedBinResize helper.
- useBinInspectorMultiActions.ts (238 LOC): bulk action callbacks
  (updateMultiCategory, updateMultiCustomProperty, updateMultiHeight,
  updateMultiClearance, updateMultiLayer) packaged as their own hook.
- useBinInspector.ts (415 LOC): main hook — state, constraints,
  single-bin actions (updateField, updateCustomProperties, moveToLayer,
  delete, rotate, etc.), composing the multi-action sub-hook.
